requirements.txt
azure.identity==1.8.0
msgraph.core==0.2.2
tabulate==0.8.9
mailjet-rest==1.3.4